requirements.txt
pandas
numpy
urlparse
Optional
Dict
RegexpTokenizer
matplotlib.pyplot
seaborn
wordcloud
LabelEncoder
Sklearn
joblib

